build(deps): Upgrade YamlDotNet to v16

+++ src/Recyclarr.Config/EnvironmentVariables/EnvironmentVariableNotDefinedException.cs
@@ -1,8 +1,8 @@
 namespace Recyclarr.Config.EnvironmentVariables;
 
-public class EnvironmentVariableNotDefinedException(int line, string envVarName) : Exception(
+public class EnvironmentVariableNotDefinedException(long line, string envVarName) : Exception(
     $"Line {line} refers to undefined environment variable {envVarName} and no default is specified.")
 {
-    public int Line { get; } = line;
+    public long Line { get; } = line;
     public string EnvironmentVariableName { get; } = envVarName;
 }
